refactor: moving zipFiles to utils as it will be used by other classes
