+++ intuitivecare/src/main/java/com/br/intuitivecare/webscraping/Scraper.java
@@ -3,9 +3,10 @@ package com.br.intuitivecare.webscraping;
 import java.io.IOException;
 import java.util.List;
 
+import com.br.intuitivecare.utils.ConfigManager;
 import com.br.intuitivecare.utils.Dialog;
+import com.br.intuitivecare.utils.ZipFiles;
 import com.br.intuitivecare.webscraping.services.DownloadFiles;
-import com.br.intuitivecare.webscraping.services.ZipFiles;
 
 public class Scraper {
 
@@ -15,7 +16,7 @@ public class Scraper {
             ZipFiles zipper = new ZipFiles();
 
             List<String> downloadedFiles = downloader.downloadAnexos();
-            zipper.createZip(downloadedFiles);
+            zipper.createZip(downloadedFiles, ConfigManager.get("scraping.zip.file"));
             Dialog.showMessage("Processo concluído!", "Sucesso");
         } catch (IOException e) {
             Dialog.showError("Erro: " + e.getMessage());
